requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-ratelimit>=4.1.0
dj-database-url>=2.1.0
whitenoise>=6.6.0
sentry-sdk>=2.43.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
razorpay>=1.4.0
Pillow>=10.0.0
gunicorn>=21.0.0
PyJWT>=2.8.0
reportlab>=4.0.0
python-magic>=0.4.27
celery>=5.3.0
redis>=5.0.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
